Reimplement using OpenSSL high-level API
This is the main rework. I initially wrote the standard functions for
SHA1, and only later refactored to extract all SHA1 specifics into macro
parameters, so that all other algorithms could be generated by the
macro.

It follows the same idea that the previous code, however it is not a
step-by-step refactoring but a rewrite from scratch.