Update compilation flags in rebar.config
There was an issue in MacOS where in 15.x the dynamic linker would give
preference to the system's cached `libboringssl.dylib`, instead of the
OpenSSL's one put in the path. For the dynamic linker to resolve the
right library correctly, the static linker would need to add strict
resolution rules by being provided more strict flags.